+++ src/handlers/ManagePluginsHandler.ts
@@ -1,23 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import * as Fs from "fs-extra";
-import * as Path from "path";
 import { ConfigurationTarget, QuickPickItem, QuickPickItemKind, Uri, WorkspaceFolder, commands, window, workspace } from "vscode";
 import { ClientManager } from "../ClientManager";
 import AutoDisposable from "../helper/AutoDisposable";
 import { VdmDialect, getDialect } from "../util/DialectUtil";
-import { getFilesFromDirRecur } from "../util/DirectoriesUtil";
 import { JarFile } from "../util/JarFile";
 import * as Util from "../util/Util";
-import { getExtensionPath } from "../util/ExtensionUtil";
 import { VDMJPrecision } from "../util/Util";
-
-type PluginType = "builtin" | "user";
-
-interface PluginSource {
-    type: PluginType;
-    jarPath: string;
-}
+import { PluginSource, VDMJExtensionsHandler } from "./VDMJExtensionsHandler";
 
 interface PluginMetadata {
     name: string;
@@ -49,13 +39,6 @@ export class ManagePluginsHandler extends AutoDisposable {
         commands.executeCommand("setContext", "vdm-vscode.managePlugins", true);
         Util.registerCommand(this._disposables, "vdm-vscode.managePlugins", (inputUri: Uri) =>
             this.handleManagePlugins(workspace.getWorkspaceFolder(inputUri))
-        );
-
-        Util.registerCommand(this._disposables, "vdm-vscode.addPluginJarFolders", () =>
-            Util.addToSettingsArray(true, "VDM plugins", "vdm-vscode.server.plugins", "searchPaths")
-        );
-        Util.registerCommand(this._disposables, "vdm-vscode.addPluginJars", () =>
-            Util.addToSettingsArray(false, "VDM plugins", "vdm-vscode.server.plugins", "searchPaths")
         );
     }
 
@@ -176,17 +159,6 @@ export class ManagePluginsHandler extends AutoDisposable {
         }
     }
 
-    public static getIncludedPluginsFolderPath(): string {
-        const pluginPath: string = Path.resolve(getExtensionPath(), "resources", "jars", "plugins");
-
-        if (!Fs.existsSync(pluginPath)) {
-            console.log("Invalid path for default plugins: " + pluginPath);
-            return "";
-        }
-
-        return pluginPath;
-    }
-
     public static getEnabledPlugins(wsFolder: WorkspaceFolder): string[] {
         const enabledPluginsConfiguration: Record<string, boolean> =
             workspace.getConfiguration("vdm-vscode.server.plugins", wsFolder.uri)?.get("enabled") ?? {};
@@ -201,125 +173,6 @@ export class ManagePluginsHandler extends AutoDisposable {
         }
 
         return enabledPluginNames;
-    }
-
-    private static resolveJarPathsFromSettings(
-        jarPaths: string[],
-        resolveFailedPaths: string[],
-        settingsLevel: string,
-        rootUri?: Uri
-    ): string[] {
-        // Resolve jar paths, flatten directories, filter duplicate jar names and inform the user
-        const visitedJarPaths: Map<string, string> = new Map<string, string>();
-        return jarPaths
-            .flatMap((originalPath: string) => {
-                let resolvedPath: string = originalPath;
-                if (rootUri && !Path.isAbsolute(originalPath)) {
-                    // Path should be relative to the project
-                    resolvedPath = Path.resolve(...[rootUri.fsPath, originalPath]);
-                }
-                if (!Fs.existsSync(resolvedPath)) {
-                    resolveFailedPaths.push(originalPath);
-                    return [];
-                }
-                return Fs.lstatSync(resolvedPath).isDirectory() ? getFilesFromDirRecur(resolvedPath, "jar") : [resolvedPath];
-            })
-            .filter((jarPath: string) => {
-                const jarName: string = Path.basename(jarPath);
-                if (!visitedJarPaths.has(jarName)) {
-                    visitedJarPaths.set(jarName, jarPath);
-                    return true;
-                }
-                window.showInformationMessage(
-                    `The plugin jar '${jarName}' is in multiple paths for the setting level ${settingsLevel}. Using the path '${visitedJarPaths.get(
-                        jarName
-                    )}'.`
-                );
-                return false;
-            });
-    }
-
-    public static getUserPluginSources(wsFolder: WorkspaceFolder): PluginSource[] {
-        // Get plugin jars specified by the user at the folder level setting - if not defined at this level then the "next up" level where it is defined is returned.
-        let folderSettings: string[] =
-            workspace.getConfiguration("vdm-vscode.server.plugins", wsFolder.uri)?.get("searchPaths") ?? ([] as string[]);
-
-        // Get plugin jars specified by the user at the user or workspace level setting - if the workspace level setting is defined then it is returned instead of the user level setting.
-        let userOrWorkspaceSettings: string[] = (workspace.getConfiguration("vdm-vscode.server.plugins")?.get("searchPaths") ??
-            []) as string[];
-
-        const resolveFailedPaths: string[] = [];
-        const jarPathsFromSettings: string[] = ManagePluginsHandler.resolveJarPathsFromSettings(
-            folderSettings,
-            resolveFailedPaths,
-            "Folder",
-            wsFolder.uri
-        );
-
-        console.log(folderSettings, userOrWorkspaceSettings);
-
-        // Determine if settings are equal, e.g. if the setting is not defined at the folder level.
-        if (
-            folderSettings.length !== userOrWorkspaceSettings.length ||
-            !folderSettings.every((ujp: string) => userOrWorkspaceSettings.find((fjp: string) => fjp === ujp))
-        ) {
-            // If the settings are not equal then merge them and in case of duplicate jar names the folder level takes precedence over the workspace/user level.
-            jarPathsFromSettings.push(
-                ...ManagePluginsHandler.resolveJarPathsFromSettings(
-                    userOrWorkspaceSettings,
-                    resolveFailedPaths,
-                    "User or Workspace"
-                ).filter((uwsPath: string) => {
-                    const uwsPathName: string = Path.basename(uwsPath);
-                    const existingJarPath: string = jarPathsFromSettings.find(
-                        (fsPath: string) => Path.basename(fsPath) === Path.basename(uwsPath)
-                    );
-                    if (existingJarPath) {
-                        window.showInformationMessage(
-                            `The library jar ${uwsPathName} has been defined on multiple setting levels. The path '${existingJarPath}' from the 'folder' level is being used.`
-                        );
-                        return false;
-                    }
-                    return true;
-                })
-            );
-        }
-
-        if (resolveFailedPaths.length > 0) {
-            const msg: string = `Unable to resolve the following VDM plugin jar/folder paths: <${resolveFailedPaths.reduce(
-                (prev, curr) => (curr += `> <${prev}`)
-            )}>. These can be changed in the settings.`;
-            console.log(msg);
-        }
-
-        return jarPathsFromSettings.map((jarPath) => ({
-            type: "user",
-            jarPath,
-        }));
-    }
-
-    public static getDefaultPluginSources(userDefinedPluginSources: PluginSource[]): PluginSource[] {
-        let includedJarsPaths: string[] = getFilesFromDirRecur(this.getIncludedPluginsFolderPath(), "jar");
-
-        if (userDefinedPluginSources.length > 0) {
-            includedJarsPaths = includedJarsPaths.filter((ijp: string) => {
-                const jarName: string = Path.basename(ijp);
-                const existingLibrarySource = userDefinedPluginSources.find((userLib) => Path.basename(userLib.jarPath) === jarName);
-                if (existingLibrarySource) {
-                    console.log(
-                        `The included library jar '${jarName}' is also defined by the user in the path '${existingLibrarySource.jarPath}'. Ignoring the version included with the extension.`
-                    );
-
-                    return false;
-                }
-                return true;
-            });
-        }
-
-        return includedJarsPaths.map((jarPath) => ({
-            type: "builtin",
-            jarPath,
-        }));
     }
 
     public static async getClasspathAdditions(wsFolder: WorkspaceFolder, dialect: VdmDialect, precision: VDMJPrecision) {
@@ -373,8 +226,7 @@ export class ManagePluginsHandler extends AutoDisposable {
         precision: VDMJPrecision,
         wsFolder: WorkspaceFolder
     ): Promise<PluginSourceMap> {
-        const userPlugins: PluginSource[] = ManagePluginsHandler.getUserPluginSources(wsFolder);
-        const plugins: PluginSource[] = [...userPlugins, ...ManagePluginsHandler.getDefaultPluginSources(userPlugins)];
+        const plugins: PluginSource[] = VDMJExtensionsHandler.getAllPluginSources(wsFolder);
 
         if (plugins.length === 0) {
             return new Map();