+++ src/util/JarFile.ts
@@ -16,6 +16,12 @@ export class JarFile {
         }
     }
 
+    public fileExists(path: string): boolean {
+        const fileObject = this._zipFile.file(path);
+
+        return fileObject === null;
+    }
+
     public async readFile(path: string): Promise<Buffer | undefined> {
         try {
             const fileObject = this._zipFile.file(path);